--- src/utils/isRoomReadyForGame.ts
@@ -1,9 +1,0 @@
-import { RoomPlayer } from './types';
-
-export function isRoomReadyForGame(
-  players: RoomPlayer[],
-  playerId: string | number
-) {
-  const isPlayerInTheRoom = players.some((p) => p.index === playerId);
-  return players.length && !isPlayerInTheRoom;
-}
