+++ src/websocket/messageHandlers/getGameMessages.ts
@@ -1,0 +1,171 @@
+import Database from '../../db/db';
+import { getShipCells } from '../../utils/getShipCells';
+import {
+  AttackStatus,
+  Game,
+  Cell,
+  Ship,
+  WebsocketCommandType,
+  AttackResult,
+} from '../../utils/types';
+import { formMessage } from '../../utils/formMessage';
+
+export function getGameMessages(
+  type: WebsocketCommandType,
+  db: Database<Game>,
+  data: any
+) {
+  const game = db.findById(data.gameId);
+
+  if (!game) {
+    return;
+  }
+
+  if (type === WebsocketCommandType.ADD_SHIPS) {
+    const updatedGame = db.updateOne(game.id, {
+      players: game.players.map((player) => {
+        if (player.playerId === data.indexPlayer) {
+          return {
+            ...player,
+            ships: data.ships.map((ship: Omit<Ship, 'life'>) => ({
+              ...ship,
+              cells: getShipCells(ship.position, ship.length, ship.direction),
+            })),
+          };
+        }
+
+        return player;
+      }),
+      firstPlayerId: game.firstPlayerId || data.indexPlayer,
+    });
+
+    const allPlayersAddedShips = updatedGame.players.every(
+      (player) => player.ships.length
+    );
+
+    if (allPlayersAddedShips) {
+      const startGameMessage = formMessage(
+        WebsocketCommandType.START_GAME,
+        updatedGame
+      );
+      const turnMessage = formMessage(WebsocketCommandType.TURN, {
+        currentPlayer: updatedGame.firstPlayerId,
+      });
+
+      return [startGameMessage, turnMessage];
+    }
+
+    return updatedGame;
+  }
+
+  if (
+    type === WebsocketCommandType.ATTACK ||
+    type === WebsocketCommandType.RANDOM_ATTACK
+  ) {
+    const { x, y, indexPlayer } = data;
+    const result: AttackResult = {
+      position: {
+        x,
+        y,
+      },
+      currentPlayer: indexPlayer,
+      status: AttackStatus.MISS,
+    };
+    const messagesForKilledShip: AttackResult[] = [];
+
+    const enemy = game.players.find(
+      (player) => player.playerId !== indexPlayer
+    );
+
+    if (!enemy || indexPlayer !== game.firstPlayerId) {
+      return;
+    }
+
+    const updatedEnemyShips = getUpdatedShips(
+      enemy.ships,
+      messagesForKilledShip,
+      result
+    );
+
+    const currentPlayer = game.players.find(
+      (player) => player.playerId === indexPlayer
+    )!;
+
+    const nextPlayerId =
+      result.status === AttackStatus.MISS ? enemy.playerId : game.firstPlayerId;
+
+    db.updateOne(game.id, {
+      firstPlayerId: nextPlayerId,
+      players: [
+        currentPlayer,
+        {
+          ...enemy,
+          ships: updatedEnemyShips,
+        },
+      ],
+    });
+
+    if (!updatedEnemyShips.length) {
+      return formMessage(WebsocketCommandType.FINISH, {
+        winPlayer: nextPlayerId,
+      });
+    }
+
+    const turnMessage = formMessage(WebsocketCommandType.TURN, {
+      currentPlayer: nextPlayerId,
+    });
+
+    if (result.status === AttackStatus.KILLED) {
+      return [
+        ...messagesForKilledShip.map((msg) =>
+          formMessage(WebsocketCommandType.ATTACK, msg)
+        ),
+        turnMessage,
+      ];
+    }
+
+    return [formMessage(WebsocketCommandType.ATTACK, result), turnMessage];
+  }
+}
+
+function getUpdatedShips(
+  enemyShips: Ship[],
+  messagesForKilledShip: AttackResult[],
+  result: AttackResult
+) {
+  const { x, y } = result.position;
+
+  return enemyShips
+    .map((ship: Ship) => {
+      const shotCell = ship.cells.find(
+        (cell: Cell) => cell.x === x && cell.y === y && !cell.isShot
+      );
+
+      if (!shotCell) {
+        return ship;
+      }
+
+      shotCell.isShot = true;
+      const isShipKilled = ship.cells.every((cell: Cell) => cell.isShot);
+
+      if (isShipKilled) {
+        ship.cells.forEach((cell: Cell) => {
+          messagesForKilledShip.push({
+            ...result,
+            position: {
+              x: cell.x,
+              y: cell.y,
+            },
+            status: AttackStatus.KILLED,
+          });
+        });
+
+        result.status = AttackStatus.KILLED;
+        return null;
+      }
+
+      result.status = AttackStatus.SHOT;
+      return ship;
+    })
+    .filter(Boolean) as Ship[];
+}
