+++ src/websocket/messageHandlers/getRegMessage.ts
